webui/requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
connectorx==0.3.1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
deltalake==0.10.0
distlib==0.3.6
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mysql-connector-python==8.0.33
numpy==1.25.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
polars==0.18.4
protobuf==3.20.3
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.17
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
virtualenv==20.23.1
watchdog==3.0.0
xlsx2csv==0.8.1
XlsxWriter==3.1.2
zipp==3.15.0
